feat: use FastApi APIRouter
