+++ cool_seq_tool/version.py
@@ -1,1 +1,1 @@
-__version__ = "0.1.7"
+__version__ = "0.1.8"
